Execution Changes

+++ StepDefinitions/CAP018_BKG_00002_RebookanalreadyexecutedAWBStepDefinitions.cs
@@ -12,7 +12,6 @@ namespace iCargoUIAutomation.StepDefinitions
         private PageObjectManager pageObjectManager;
         private homePage hp;
         private MaintainBookingPage mbp;
-        private CreateShipmentPage csp;
         string awb = "";
         string flightOrigin = "";
         string flightDestination = "";
@@ -34,32 +33,43 @@ namespace iCargoUIAutomation.StepDefinitions
         [Then(@"User deletes the flight details and adds new flight details")]
         public void ThenUserDeletesTheFlightDetailsAndAddsNewFlightDetails()
         {
-            mbp.DeleteAddFlights();
-        }
-
-        //[Then(@"User enters new Carrier details with Origin ""([^""]*)"", Destination ""([^""]*)"", Flight No ""([^""]*)"", Flight Date ""([^""]*)"", Pieces ""([^""]*)"", Weight ""([^""]*)""")]
-        //public void ThenUserEntersNewCarrierDetailsWithOriginDestinationFlightNoFlightDatePiecesWeight(string flightOrigin, string flightDestination, string flightNumber, string flightDate, string flightPieces, string flightWeight)
-        //{
-        //    this.flightOrigin = flightOrigin;
-        //    this.flightDestination = flightDestination;
-        //    this.flightNumber = flightNumber;
-        //    this.flightDate = flightDate;
-        //    this.flightPieces = flightPieces;
-        //    this.flightWeight = flightWeight;
-        //    mbp.addNewFlightDetails(flightOrigin,flightDestination, flightNumber, flightDate, flightPieces, flightWeight);
-        //}
+            if (ScenarioContext.Current["Execute"] == "true")
+            {
+                mbp.DeleteAddFlights();
+            }
+            else
+            {
+                ScenarioContext.Current.Pending();
+            }
+            
+        }        
 
         [Then(@"User selects new carrier details")]
         public void ThenUserSelectsNewCarrierDetails()
         {
-            mbp.addNewFlightDetails();
+            if (ScenarioContext.Current["Execute"] == "true")
+            {
+                mbp.AddNewFlightDetails();
+            }
+            else
+            {
+                ScenarioContext.Current.Pending();
+            }
+            
         }
 
 
         [Then(@"User clicks on Save button to save new flight details")]
         public void ThenUserClicksOnSaveButtonToSaveNewFlightDetails()
         {
-            mbp.clickOnSaveButtonToSaveNewFlightDetails();
+            if (ScenarioContext.Current["Execute"] == "true")
+            {
+                mbp.clickOnSaveButtonToSaveNewFlightDetails();
+            }
+            else
+            {
+                ScenarioContext.Current.Pending();
+            }            
         }
 
 
@@ -67,15 +77,15 @@ namespace iCargoUIAutomation.StepDefinitions
         [Then(@"User captures the irregularity details")]
         public void ThenUserCapturesTheIrregularityDetails()
         {
-            mbp.CaptureIrregularity();
+            if (ScenarioContext.Current["Execute"] == "true")
+            {
+                mbp.CaptureIrregularity();
+            }
+            else
+            {
+                ScenarioContext.Current.Pending();
+            }            
         }
-
-        [Then(@"User enters already executed AWB number")]
-        public void ThenUserEntersAlreadyExecutedAWBNumber()
-        {
-           csp.EnterAWBNumberForMaintainBooking();
-        }
-
 
 
     }
